+++ frontend/.next/server/webpack-runtime.js
@@ -125,7 +125,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("224494148576a732")
+/******/ 		__webpack_require__.h = () => ("f81a8e171e5666b6")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */