+++ frontend/.next/static/chunks/webpack.js
@@ -163,7 +163,7 @@
 /******/ 		// This function allow to reference async chunks
 /******/ 		__webpack_require__.u = function(chunkId) {
 /******/ 			// return url for filenames based on template
-/******/ 			return "static/chunks/" + chunkId + ".js";
+/******/ 			return undefined;
 /******/ 		};
 /******/ 	}();
 /******/ 	
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "566a42f6e5f876ca"; }
+/******/ 		__webpack_require__.h = function() { return "308ad2ff2c2dbf15"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */